requirements.txt
anthropic==0.75.0
python-telegram-bot==22.5
boto3==1.42.15
python-dotenv==1.2.1
pydantic==2.12.5
pydantic-settings==2.12.0
google-genai==1.56.0
Pillow==12.0.0
openai==2.14.0
supabase==2.27.0
flask==3.1.2
gunicorn==23.0.0

